requirements.txt
streamlit
torch
transformers
diffusers
accelerate
safetensors
imageio[ffmpeg]
Pillow
google-generativeai
